+++ src/libstd/old_path/windows.rs
@@ -1124,7 +1124,7 @@ mod tests {
     use clone::Clone;
     use iter::IteratorExt;
     use option::Option::{self, Some, None};
-    use path::GenericPath;
+    use old_path::GenericPath;
     use slice::{AsSlice, SliceExt};
     use str::Str;
     use string::ToString;
