+++ src/test/compile-fail/range-3.rs
