+++ src/test/compile-fail/range-4.rs
