+++ src/test/debuginfo/associated-types.rs
